requirements.txt
anthropic>=0.34.0
litellm>=1.42.14
playwright>=1.45.0
requests>=2.31.0
tweepy>=4.14.0
pandas>=2.2.0
torch>=2.2.0
transformers>=4.42.0
openai>=1.35.0
python-dotenv>=1.0.1
pyyaml>=6.0.1

